Implement lazy lookup of references instead of maintaining explicit links.
The sqlite db is so fast that lazy lookups are just as fast, and allows further
reducing complexity in the data structure. Don't bother tracking references and
trying to figure out when to update which types, etc. Just update the search
index and search for references on the fly! Also speeds up initial parsing.